@@ -58,36 +49,36 @@
   },
   "devDependencies": {
     "@types/colors": "^1.1.3",
-    "@types/debug": "0.0.29",
-    "@types/express": "^4.0.35",
+    "@types/debug": "0.0.30",
+    "@types/express": "^4.11.0",
     "@types/jasmine": "2.5.45",
-    "@types/lodash": "^4.14.63",
-    "@types/mongodb": "^2.2.0",
-    "@types/node": "^8.0.46",
-    "@types/supertest": "^2.0.0",
+    "@types/lodash": "^4.14.92",
+    "@types/mongodb": "^2.2.18",
+    "@types/node": "^9.3.0",
+    "@types/supertest": "^2.0.4",
     "@types/uuid": "^3.4.3",
-    "body-parser": "^1.17.1",
-    "helmet": "^3.6.0",
-    "http-server": "^0.9.0",
+    "body-parser": "^1.18.2",
+    "helmet": "^3.9.0",
+    "http-server": "^0.10.0",
     "istanbul": "^0.4.5",
-    "jasmine": "^2.6.0",
-    "jasmine-spec-reporter": "^3.3.0",
-    "nodemon": "^1.11.0",
-    "nsp": "^2.6.3",
-    "source-map-support": "^0.4.14",
+    "jasmine": "^2.8.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "nodemon": "^1.14.9",
+    "nsp": "^3.1.0",
+    "source-map-support": "^0.5.0",
     "supertest": "^3.0.0",
-    "tslint": "^5.1.0",
+    "tslint": "^5.8.0",
     "typedoc": "^0.9.0",
-    "typescript": "^2.3.1"
+    "typescript": "^2.6.2"
   },
   "dependencies": {
     "bson-ext": "^1.0.5",
     "co": "^4.6.0",
     "colors": "^1.1.2",
-    "debug": "^2.6.6",
-    "express": "^4.15.2",
+    "debug": "^3.1.0",
+    "express": "^4.16.2",
     "lodash": "^4.17.4",
-    "mongodb": "^2.2.26",
+    "mongodb": "^2.2.33",
     "reflect-metadata": "^0.1.10",
     "uuid": "^3.1.0"
   }
