@@ -1,5 +1,5 @@
 import {DisplayProcessor, SpecReporter} from 'jasmine-spec-reporter';
-import 'source-map-support/register';
+import 'source-map-support';
 import SuiteInfo = jasmine.SuiteInfo;
 
 class CustomProcessor extends DisplayProcessor {
