@@ -1,7 +1,7 @@
+import * as path from 'path';
 import {testMongoDbUrl, testSapi} from '../../spec/helpers/sakuraapi';
 import {SakuraMongoDbConnection} from '../core/sakura-mongo-db-connection';
 import {SakuraApiConfig} from './sakura-api-config';
-import path = require('path');
 
 describe('sakura-api-config', () => {
 
