@@ -3,4 +3,4 @@ export * from './errors';
 export * from './json';
 export * from './model';
 export * from './private';
-export * from './sakura-api-model';
+export * from './sapi-model-mixin';
