@@ -1,3 +1,3 @@
 export * from './routable';
 export * from './route';
-export * from './sakura-api-routable';
+export * from './sapi-routable-mixin';
