@@ -1,22 +1,44 @@
-import * as express from 'express';
-import {NextFunction, Request, Response} from 'express';
-import * as request from 'supertest';
-import {testSapi, testUrl} from '../../../spec/helpers/sakuraapi';
-import {Db, Json, Model, SakuraApiModel} from '../@model';
-import {IRoutableLocals, Routable, routableSymbols, Route} from './';
-import {SakuraApiRoutable} from './sakura-api-routable';
+import {
+  NextFunction,
+  Request,
+  Response
+}                          from 'express';
+import * as request        from 'supertest';
+import {
+  testSapi,
+  testUrl
+}                          from '../../../spec/helpers/sakuraapi';
+import {
+  Db,
+  Json,
+  Model
+}                          from '../@model';
+import {SapiModelMixin}    from '../@model/sapi-model-mixin';
+import {
+  AuthenticatorPlugin,
+  AuthenticatorPluginResult,
+  IAuthenticator,
+  IAuthenticatorConstructor
+}                          from '../plugins';
+import {
+  IRoutableLocals,
+  Routable,
+  routableSymbols,
+  Route
+}                          from './';
+import {SapiRoutableMixin} from './sapi-routable-mixin';
 
 describe('core/Route', () => {
   @Routable({
     baseUrl: 'testCoreRoute',
     blackList: ['someBlacklistedMethod']
   })
-  class TestCoreRoute extends SakuraApiRoutable {
+  class TestCoreRoute extends SapiRoutableMixin() {
     @Route({
       method: 'get',
       path: '/'
     })
-    someMethod(req: express.Request, res: express.Response) {
+    someMethod(req: Request, res: Response) {
       res
         .status(200)
         .send({someMethodCalled: true});
@@ -26,7 +48,7 @@ describe('core/Route', () => {
       method: 'post',
       path: 'someOtherMethod/'
     })
-    someOtherMethod(req: express.Request, res: express.Response) {
+    someOtherMethod(req: Request, res: Response) {
       res
         .status(200)
         .send({someOtherMethodCalled: true});
@@ -36,7 +58,7 @@ describe('core/Route', () => {
       method: 'post',
       path: 'someBlacklistedMethod/'
     })
-    someBlacklistedMethod(req: express.Request, res: express.Response) {
+    someBlacklistedMethod(req: Request, res: Response) {
       res
         .status(200)
         .send({someOtherMethodCalled: true});
@@ -253,7 +275,7 @@ describe('core/Route', () => {
         db: 'userDb'
       }
     })
-    class AfterHandlerTestModel extends SakuraApiModel {
+    class AfterHandlerTestModel extends SapiModelMixin() {
       @Db() @Json()
       firstName = 'George';
 
@@ -344,6 +366,94 @@ describe('core/Route', () => {
         })
         .then(done)
         .catch(done.fail);
+    });
+  });
+
+  describe('authenticators', () => {
+    @AuthenticatorPlugin()
+    class RoutableAuthenticator implements IAuthenticator, IAuthenticatorConstructor {
+      async authenticate(req: Request, res: Response): Promise<AuthenticatorPluginResult> {
+        return {data: {}, status: 200, success: true};
+      }
+    }
+
+    @AuthenticatorPlugin()
+    class RouteAuthenticator implements IAuthenticator, IAuthenticatorConstructor {
+      async authenticate(req: Request, res: Response): Promise<AuthenticatorPluginResult> {
+        return {data: {}, status: 200, success: true};
+      }
+    }
+
+    it('injects route authenticators array into @Routable metadata', () => {
+
+      @Routable()
+      class TestRoutable extends SapiRoutableMixin() {
+
+        @Route({
+          authenticator: [RouteAuthenticator]
+        })
+        testHandler() {
+
+        }
+      }
+
+      const sapi = testSapi({
+        routables: [TestRoutable]
+      });
+
+      const authenticators = Reflect.getMetadata('authenticators.testHandler', new TestRoutable());
+
+      expect(authenticators).toBeDefined();
+      expect(Array.isArray(authenticators)).toBeTruthy();
+      expect(authenticators[0]).toBe(RouteAuthenticator);
+
+    });
+
+    it('injects route authenticator as array into @Routable metadata', () => {
+
+      @Routable()
+      class TestRoutable extends SapiRoutableMixin() {
+
+        @Route({
+          authenticator: RouteAuthenticator
+        })
+        testHandler() {
+
+        }
+      }
+
+      const sapi = testSapi({
+        routables: [TestRoutable]
+      });
+
+      const authenticators = Reflect.getMetadata('authenticators.testHandler', new TestRoutable());
+
+      expect(authenticators).toBeDefined();
+      expect(Array.isArray(authenticators)).toBeTruthy();
+      expect(authenticators[0]).toBe(RouteAuthenticator);
+
+    });
+
+    it('injects empty with no authenticators into @Routable metadata', () => {
+
+      @Routable()
+      class TestRoutable extends SapiRoutableMixin() {
+
+        @Route()
+        testHandler() {
+        }
+      }
+
+      const sapi = testSapi({
+        routables: [TestRoutable]
+      });
+
+      const authenticators = Reflect.getMetadata('authenticators.testHandler', new TestRoutable());
+
+      expect(authenticators).toBeDefined();
+      expect(Array.isArray(authenticators)).toBeTruthy();
+      expect(authenticators.length).toBe(0);
+
     });
   });
 });
